SHA-256 collision
While it is theoretically possible for two different inputs to produce the same hash (this would be called a collision), SHA-256 is designed in such a way that finding such a collision is practically impossible with current computational resources. The space of possible hashes (2^256) is so large that the probability of randomly encountering a collision is astronomically low.